# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Web API Express",
      "description": "My WebApi using Node",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:4000",
    "basePath": "/",
    "tags": [
      {
        "name": "Players",
        "description": "API de jugadores en la aplicación"
      },
      {
        "name": "Usuario",
        "description": "API de usuarios de la BD"
      }
    ],
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/api/getPlayers": {
        "get": {
          "tags": [
            "Players"
          ],
          "summary": "Obtiene todos los jugadores",
          "description": "Regresa todos los jugadores",
          "operationId": "getPlayers",
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Player"
              }
            }
          }
        }
      },
      "/api/getPlayer/{id}": {
        "get": {
          "tags": [
            "Players"
          ],
          "summary": "Encuentra un jugador por id",
          "description": "Regresa el jugador con el id solicitado",
          "operationId": "getPlayer",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Id del jugador a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Player"
              }
            }
          }
        }
      },
      "/api/addPlayer": {
        "post": {
          "tags": [
            "Players"
          ],
          "summary": "Agrega un nuevo jugador",
          "description": "Agrega un jugador a la base de datos",
          "operationId": "addPlayer",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Nuevo jugador",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Player"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Player"
              }
            },
            "405": {
              "description": "Datos inválidos"
            }
          }
        }
      },
      "/api/updatePlayer/{id}": {
        "put": {
          "tags": [
            "Players"
          ],
          "summary": "Actualiza un nuevo jugador",
          "description": "Actualiza un jugador a la base de datos",
          "operationId": "updatePlayer",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Id del jugador a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Datos nuevos jugador",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Player"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Player"
              }
            },
            "400": {
              "description": "Jugador Inválido"
            },
            "404": {
              "description": "Jugador no encontrado"
            }
          }
        }
      },
      "/api/deletePlayer/{id}": {
        "delete": {
          "tags": [
            "Players"
          ],
          "summary": "Borra un jugador",
          "description": "Borra un jugador de la base de datos",
          "operationId": "deletePlayer",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Id del jugador a borrar",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Player"
              }
            },
            "400": {
              "description": "Jugador Inválido"
            },
            "404": {
              "description": "Jugador no encontrado"
            }
          }
        }
      },
      "/api/getUsuarios": {
        "get": {
          "tags": [
            "Usuario"
          ],
          "summary": "Obtiene todos los usuarios",
          "description": "Regresa todos los usuarios de la tabla Usuario",
          "operationId": "get",
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          }
        }
      },
      "/api/getUsuario/{correo}": {
        "get": {
          "tags": [
            "Usuario"
          ],
          "summary": "Encuentra un jugador por id",
          "description": "Regresa el jugador con el id solicitado",
          "operationId": "getUsuario",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "correo",
              "description": "Id del jugador a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          }
        }
      },
      "/api/addUsuarios": {
        "post": {
          "tags": [
            "Usuario"
          ],
          "summary": "Agrega un nuevo usuario",
          "description": "Agrega un usuario a la base de datos",
          "operationId": "post",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Nuevo usuario",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "405": {
              "description": "Datos inválidos"
            }
          }
        }
      },
      "/api/updateUsuarioImg/{idUsuarios}": {
        "put": {
          "tags": [
            "Usuario"
          ],
          "summary": "Agrega un nuevo usuario",
          "description": "Agrega un usuario a la base de datos",
          "operationId": "updateUsuarioImg",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "idUsuarios",
              "description": "Id del jugador a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Datos nuevos jugador",
              "required": true,
              "schema": {
                "$ref": "#/definitions/UsuarioImg"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "405": {
              "description": "Datos inválidos"
            }
          }
        }
      },
      "/api/getMedallas/{Usuarios_idUsuarios}": {
        "get": {
          "tags": [
            "Medallas"
          ],
          "summary": "Encuentra un jugador por id",
          "description": "Regresa el jugador con el id solicitado",
          "operationId": "getMedallas",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "Usuarios_idUsuarios",
              "description": "Id del jugador a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Medallas"
              }
            }
          }
        }
      },
      "/api/getLogros/{Usuarios_idUsuarios}": {
        "get": {
          "tags": [
            "Medallas"
          ],
          "summary": "Encuentra un jugador por id",
          "description": "Regresa el jugador con el id solicitado",
          "operationId": "getLogros",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "Usuarios_idUsuarios",
              "description": "Id del jugador a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "schema": {
                "$ref": "#/definitions/Medallas"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Player": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "namePlayer": {
            "type": "string"
          },
          "avatarPlayer": {
            "type": "string"
          },
          "points": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Medallas": {
        "type": "object",
        "properties": {
          "Usuarios_idUsuarios": {
            "type": "string"
          },
          "Medallas_idMedallas": {
            "type": "string"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "idUsuarios": {
            "type": "integer",
            "format": "int64"
          },
          "correo": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "apellidoP": {
            "type": "string"
          },
          "apellidoM": {
            "type": "string"
          },
          "tipoUsuario": {
            "type": "integer",
            "format": "int64"
          },
          "foto": {
            "type":"string"
          }
        }
      },
      "UsuarioImg": {
        "type": "object",
        "properties": {
          "idUsuarios": {
            "type": "integer",
            "format": "int64"
          },
          "foto": {
            "type": "string"
          }
        }
      }
    }
  }